Rename all occurrences of (gc) works to work
