+++ src/plan/barriers.rs
@@ -1,5 +1,5 @@
 use crate::policy::space::Space;
-use crate::scheduler::gc_works::*;
+use crate::scheduler::gc_work::*;
 use crate::scheduler::WorkBucketStage;
 use crate::util::*;
 use crate::MMTK;
