+++ src/plan/semispace/mod.rs
@@ -1,4 +1,4 @@
-pub(super) mod gc_works;
+pub(super) mod gc_work;
 pub(super) mod global;
 pub(super) mod mutator;
