+++ src/scheduler/gc_work.rs
@@ -47,7 +47,7 @@ impl<P: Plan, W: CopyContext + WorkerLocal> GCWork<P::VM> for Prepare<P, W> {
                 .add(PrepareMutator::<P::VM>::new(mutator));
         }
         for w in &mmtk.scheduler.worker_group().workers {
-            w.local_works.add(PrepareCollector::<W>::new());
+            w.local_work.add(PrepareCollector::<W>::new());
         }
     }
 }
@@ -115,7 +115,7 @@ impl<P: Plan, W: CopyContext + WorkerLocal> GCWork<P::VM> for Release<P, W> {
                 .add(ReleaseMutator::<P::VM>::new(mutator));
         }
         for w in &mmtk.scheduler.worker_group().workers {
-            w.local_works.add(ReleaseCollector::<W>(PhantomData));
+            w.local_work.add(ReleaseCollector::<W>(PhantomData));
         }
         // TODO: Process weak references properly
         mmtk.reference_processors.clear();
@@ -381,9 +381,9 @@ pub trait ProcessEdgesWork:
         if Self::SCAN_OBJECTS_IMMEDIATELY {
             // We execute this `scan_objects_work` immediately.
             // This is expected to be a useful optimization because,
-            // say for _pmd_ with 200M heap, we're likely to have 50000~60000 `ScanObjects` works
+            // say for _pmd_ with 200M heap, we're likely to have 50000~60000 `ScanObjects` work
             // being dispatched (similar amount to `ProcessEdgesWork`).
-            // Executing these works now can remarkably reduce the global synchronization time.
+            // Executing these work now can remarkably reduce the global synchronization time.
             self.worker().do_work(scan_objects_work);
         } else {
             self.mmtk.scheduler.work_buckets[WorkBucketStage::Closure].add(scan_objects_work);
