+++ src/scheduler/work.rs
@@ -19,7 +19,7 @@ pub trait Work<C: Context>: 'static + Send + Sync {
 /// A special kind of work that will execute on the coorddinator (i.e. controller) thread
 ///
 /// The coorddinator thread holds the global monitor lock when executing `CoordinatorWork`s.
-/// So, directly adding new works to any buckets will cause dead lock.
+/// So, directly adding new work to any buckets will cause dead lock.
 /// For this case, use `WorkBucket::add_with_priority_unsync` instead.
 pub trait CoordinatorWork<C: Context>: 'static + Send + Sync + Work<C> {}
