+++ src/scheduler/worker.rs
@@ -50,7 +50,7 @@ pub struct Worker<C: Context> {
     pub parked: AtomicBool,
     scheduler: Arc<Scheduler<C>>,
     local: WorkerLocalPtr,
-    pub local_works: WorkBucket<C>,
+    pub local_work: WorkBucket<C>,
     pub sender: Sender<CoordinatorMessage<C>>,
     pub stat: WorkerLocalStat,
     context: Option<&'static C>,
@@ -71,7 +71,7 @@ impl<C: Context> Worker<C> {
             ordinal,
             parked: AtomicBool::new(true),
             local: WorkerLocalPtr::UNINITIALIZED,
-            local_works: WorkBucket::new(true, scheduler.worker_monitor.clone()),
+            local_work: WorkBucket::new(true, scheduler.worker_monitor.clone()),
             sender: scheduler.channel.0.clone(),
             scheduler,
             stat: Default::default(),
