+++ src/vm/scanning.rs
@@ -1,5 +1,5 @@
 use crate::plan::{Mutator, TransitiveClosure};
-use crate::scheduler::gc_works::ProcessEdgesWork;
+use crate::scheduler::gc_work::ProcessEdgesWork;
 use crate::scheduler::GCWorker;
 use crate::util::ObjectReference;
 use crate::util::OpaquePointer;
